Move firebase to services and adds contexts folder.
