--- src/components/Firebase/index.js
@@ -1,6 +1,0 @@
-import FirebaseContext, { withFirebase } from './context';
-import Firebase from './firebase';
-
-export default Firebase;
-
-export { FirebaseContext, withFirebase };
