--- src/components/Session/context.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const AuthUserContext = React.createContext(null);
-
-export default AuthUserContext;
