--- src/components/Session/index.js
@@ -1,4 +1,0 @@
-import AuthUserContext from './context';
-import withAuthorization from './withAuthorization';
-
-export { AuthUserContext, withAuthorization };
