--- src/components/Session/withAuthorization.js
@@ -1,39 +1,0 @@
-import React from 'react';
-import { withRouter } from 'react-router-dom';
-import { compose } from 'recompose';
-
-import AuthUserContext from './context';
-import { withFirebase } from '../Firebase';
-import * as ROUTES from '../../constants/routes';
-
-// eslint-disable-next-line no-unused-vars
-const withAuthorization = condition => Component => props => {
-  class WithAuthorization extends React.Component {
-    componentDidMount() {
-      const { firebase, history } = this.props;
-      this.listener = firebase.auth.onAuthStateChanged(authUser => {
-        if (!condition(authUser)) {
-          history.push(ROUTES.SIGN_IN);
-        }
-      });
-    }
-
-    componentWillUnmount() {
-      this.listener();
-    }
-
-    render() {
-      return (
-        <AuthUserContext.Consumer>
-          {authUser =>
-            condition(authUser) ? <Component {...this.props} /> : null
-          }
-        </AuthUserContext.Consumer>
-      );
-    }
-  }
-
-  return compose(withRouter, withFirebase)(WithAuthorization);
-};
-
-export default withAuthorization;
