+++ src/contexts/index.js
@@ -1,0 +1,4 @@
+import React from 'react';
+
+export const AuthUserContext = React.createContext(null);
+export const FirebaseContext = React.createContext(null);
