+++ src/index.js
@@ -2,7 +2,8 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 
 import App from './components/App';
-import Firebase, { FirebaseContext } from './components/Firebase';
+import Firebase from './services/Firebase';
+import { FirebaseContext } from './contexts';
 
 import * as serviceWorker from './serviceWorker';
