+++ src/services/Firebase/index.js
@@ -1,0 +1,6 @@
+import withFirebase from './withFirebase';
+import Firebase from './firebase';
+
+export default Firebase;
+
+export { withFirebase };
